server/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
screenpoint==0.1.1
pyscreenshot==1.0
flask==1.1.1
flask-cors==3.0.8
gunicorn==20.0.4
scikit-image==0.14.0
Pillow==6.2.2
torch==1.3.1+cpu 
torchvision==0.4.2+cpu
requests==2.25.1
scipy==1.5.4
numpy==1.15.2